feat(trait): add emailprovider to match async architecture
* Also update mod to fix all isues in providers folder
